Merge

+++ src/services/combat/CombatService.java
@@ -66,7 +66,9 @@ import services.sui.SUIService.MessageBoxType;
 import services.sui.SUIWindow;
 import services.sui.SUIWindow.SUICallback;
 import services.sui.SUIWindow.Trigger;
+import tools.DevLog;
 import main.NGECore;
+import engine.clients.Client;
 import engine.resources.common.CRC;
 import engine.resources.objects.SWGObject;
 import engine.resources.scene.Point3D;
@@ -152,15 +154,17 @@ public class CombatService implements INetworkDispatch {
 
 					@Override
 					public void run() {
-						
-						if(command.getDelayAttackParticle().length() > 0)
-							target.notifyObservers(new PlayClientEffectLocMessage(command.getDelayAttackParticle(), target.getPlanet().getName(), target.getWorldPosition()), true);
-
-						if(command.getAttackType() == 1)
-							doSingleTargetCombat(attacker, target, weapon, command, actionCounter);
-						else if(command.getAttackType() == 0 || command.getAttackType() == 2 || command.getAttackType() == 3)
-							doAreaCombat(attacker, targetPos, weapon, command, actionCounter, targetParent);
-						
+						try {
+							if(command.getDelayAttackParticle().length() > 0)
+								target.notifyObservers(new PlayClientEffectLocMessage(command.getDelayAttackParticle(), target.getPlanet().getName(), target.getWorldPosition()), true);
+	
+							if(command.getAttackType() == 1)
+								doSingleTargetCombat(attacker, target, weapon, command, actionCounter);
+							else if(command.getAttackType() == 0 || command.getAttackType() == 2 || command.getAttackType() == 3)
+								doAreaCombat(attacker, targetPos, weapon, command, actionCounter, targetParent);
+						} catch (Exception e) {
+							e.printStackTrace();
+						}
 					}
 					
 				}, 0, (long) (command.getDelayAttackInterval() * 1000), TimeUnit.MILLISECONDS);
@@ -169,7 +173,11 @@ public class CombatService implements INetworkDispatch {
 
 					@Override
 					public void run() {
-						task.cancel(true);
+						try {
+							task.cancel(true);
+						} catch (Exception e) {
+							e.printStackTrace();
+						}
 					}
 					
 				}, (long) ((command.getDelayAttackInterval() * 1000) * command.getDelayAttackLoops()), TimeUnit.MILLISECONDS);
@@ -203,7 +211,24 @@ public class CombatService implements INetworkDispatch {
 					creature.removeDefender(defender);
 				}
 				
-				if(((CreatureObject) target).getPlayerObject() == null) target.setKiller(attacker);
+				if(((CreatureObject) target).getPlayerObject() == null) 
+				{
+					target.setKiller(attacker);
+					
+					scheduler.schedule(new Runnable() 
+					{
+						@Override
+						public void run() 
+						{
+							try {
+								core.objectService.destroyObject(target);
+							} catch (Exception e) {
+								e.printStackTrace();
+							}
+						}
+						
+					}, 2, TimeUnit.MINUTES);
+				}
 			}
 		}
 		else if(target instanceof TangibleObject)
@@ -289,17 +314,29 @@ public class CombatService implements INetworkDispatch {
 	}
 	
 	private void applyDamage(CreatureObject attacker, TangibleObject target, int damage) {
+
 		target.setConditionDamage(target.getConditionDamage() + damage);
 		
 		if (target.getOption(Options.MOUNT)) {
 			core.mountService.damage((CreatureObject) target);
+		}
+
+		if (target instanceof CreatureObject){
+			CreatureObject targetCreature = (CreatureObject)target;
+			if (targetCreature.getCalledPet()!=null) {
+				AIActor petActor = (AIActor) targetCreature.getCalledPet().getAttachment("AI");
+				if (petActor!=null){
+					petActor.addDefender(attacker);
+					DevLog.debugout("Charon", "Pet AI", "applyDamage addDefender");
+				}
+			}
 		}
 		
 		DamageTaken event = events.new DamageTaken();
 		event.attacker = attacker;
 		event.damage = damage;
 		target.getEventBus().publish(event);
-		
+		System.out.println("APPLY DAMAGE2");
 		attacker.setTefTime(300000);
 	}
 
@@ -588,7 +625,7 @@ public class CombatService implements INetworkDispatch {
 	}
 	
 	private float calculateDamage(CreatureObject attacker, CreatureObject target, WeaponObject weapon, CombatCommand command) {
-		
+
 		if(target.getBuffByName("me_stasis_self_1") != null || target.getBuffByName("me_stasis_1") != null)
 			return 0;
 
@@ -637,14 +674,18 @@ public class CombatService implements INetworkDispatch {
 		if(target.getSkillMod("combat_divide_damage_dealt") != null) {
 			rawDamage *= (1 - (target.getSkillMod("combat_divide_damage_dealt").getBase() / 100));			
 		}
-				
+		
+		
+		//Hook in here to agitate called pet
+		
+		
 		return rawDamage;
 		
 	}
 
 	
 	private float calculateDamage(CreatureObject attacker, TangibleObject target, WeaponObject weapon, CombatCommand command) {
-		
+
 		float rawDamage = command.getAddedDamage();
 		
 		if(command.getPercentFromWeapon() > 0 && weapon != attacker.getSlottedObject("default_weapon")) {
@@ -801,8 +842,23 @@ public class CombatService implements INetworkDispatch {
 		if(target.getHealth() - damage <= 0 && target.getSlottedObject("ghost") != null) {
 			
 			if(target.hasBuff("incapWeaken")) {
-				deathblowPlayer(attacker, target);
+				if (!attacker.isPlayer()){
+					AIActor aiActor = (AIActor)attacker.getAttachment("AI");						
+					if (aiActor.getMobileTemplate().isDeathblow())
+						deathblowPlayer(attacker, target);
+				} else
+				{
+					deathblowPlayer(attacker, target);
+				}
 				return;
+			}
+			
+			if (!attacker.isPlayer()){
+				AIActor aiActor = (AIActor)attacker.getAttachment("AI");	
+				if (aiActor.getMobileTemplate().isDeathblow()){
+					deathblowPlayer(attacker, target);
+					return;
+				}
 			}
 			
 			synchronized(target.getMutex()) {
@@ -813,19 +869,23 @@ public class CombatService implements INetworkDispatch {
 				target.setHealth(1);
 				target.setPosture(Posture.Incapacitated);
 				target.setTurnRadius(0);
-				target.setSpeedMultiplierBase(0);		
+				target.setSpeedMultiplierBase(0);	
+				
 			}
+
 			ScheduledFuture<?> incapTask = scheduler.schedule(() -> {
 				
 				synchronized(target.getMutex()) {
-
-					if(target.getPosture() != 13)
-						return;
-					
-					target.setPosture(Posture.Upright);
-					target.setTurnRadius(1);
-					target.setSpeedMultiplierBase(1);
-				
+					try {
+						if(target.getPosture() != 13)
+							return;
+						
+						target.setPosture(Posture.Upright);
+						target.setTurnRadius(1);
+						target.setSpeedMultiplierBase(1);
+					} catch (Exception e) {
+						e.printStackTrace();
+					}
 				}
 			
 			}, target.getIncapTimer(), TimeUnit.SECONDS);
@@ -848,10 +908,26 @@ public class CombatService implements INetworkDispatch {
 		synchronized(target.getMutex()) {
 			target.setHealth(target.getHealth() - damage);
 		}
-		DamageTaken event = events.new DamageTaken();
-		event.attacker = attacker;
-		event.damage = damage;
-		target.getEventBus().publish(event);
+		
+
+		if (target instanceof CreatureObject){
+			CreatureObject targetCreature = (CreatureObject)target;
+			if (targetCreature.getCalledPet()!=null) {
+				AIActor petActor = (AIActor) targetCreature.getCalledPet().getAttachment("AI");
+				if (petActor!=null){
+					petActor.addDefender(attacker);
+					DevLog.debugout("Charon", "Pet AI", "applyDamage addDefender");
+				}
+			}
+		}
+		
+		
+		if (target.getPosture()!=13){
+			DamageTaken event = events.new DamageTaken();
+			event.attacker = attacker;
+			event.damage = damage;
+			target.getEventBus().publish(event);
+		}
 	}
 	
 	public void doDrainHeal(CreatureObject receiver, int drainAmount) {
@@ -929,7 +1005,7 @@ public class CombatService implements INetworkDispatch {
 		if(success && !applySpecialCost(healer, weapon, command))
 			success = false;
 
-		if(!success) {
+		if(!success && healer.getClient() != null) {
 			IoSession session = healer.getClient().getSession();
 			CommandEnqueueRemove commandRemove = new CommandEnqueueRemove(healer.getObjectId(), actionCounter);
 			session.write(new ObjControllerMessage(0x0B, commandRemove).serialize());
@@ -1027,7 +1103,7 @@ public class CombatService implements INetworkDispatch {
 					
 					attacker.sendSystemMessage(OutOfBand.ProsePackage("@bounty_hunter:bounty_success_hunter", mission.getCreditReward(), "TT", target.getCustomName(), "TO", target.getObjectID()), DisplayType.Broadcast);
 					
-					target.sendSystemMessage(OutOfBand.ProsePackage("@bounty_hunter:bounty_success_hunter", mission.getCreditReward(), "TT", attacker.getCustomName()), DisplayType.Broadcast);
+					target.sendSystemMessage(OutOfBand.ProsePackage("@bounty_hunter:bounty_success_target", mission.getCreditReward(), "TT", attacker.getCustomName()), DisplayType.Broadcast);
 					
 					core.missionService.handleMissionComplete(attacker, mission);
 				}
@@ -1229,14 +1305,16 @@ public class CombatService implements INetworkDispatch {
 
 			@Override
 			public void run() {
-
-				if(dot.getRemainingDuration() <= 0) {
-					target.removeDot(dot);
-					dot.getTask().cancel(true);
+				try {
+					if(dot.getRemainingDuration() <= 0) {
+						target.removeDot(dot);
+						dot.getTask().cancel(true);
+					}
+					
+					doDotDamageTick(attacker, target, command, dot);
+				} catch (Exception e) {
+					e.printStackTrace();
 				}
-				
-				doDotDamageTick(attacker, target, command, dot);
-				
 			}
 			
 		}, 10, 2000, TimeUnit.MILLISECONDS);
@@ -1302,7 +1380,20 @@ public class CombatService implements INetworkDispatch {
 		if(!applySpecialCost(creature, weapon, command))
 			success = false;
 		
-		if(!success) {
+		// Quick Hotfix for AI NPCs buffing themselves
+		Client client = null;
+		if(creature.getClient() == null) {
+			if (creature.getAttachment("AI")!=null){
+				AIActor aiActor = (AIActor) creature.getAttachment("AI");
+				CreatureObject target = aiActor.getFollowObject();
+				if (target.getClient()!=null)
+					client = target.getClient(); // Use player's client reference that is in combat with that NPC
+			}
+		} else {
+			client = creature.getClient();
+		}
+		
+		if(!success && client != null) {
 			IoSession session = creature.getClient().getSession();
 			CommandEnqueueRemove commandRemove = new CommandEnqueueRemove(creature.getObjectId(), actionCounter);
 			session.write(new ObjControllerMessage(0x0B, commandRemove).serialize());
@@ -1311,6 +1402,9 @@ public class CombatService implements INetworkDispatch {
 			return;
 		}
 		
+		if (client == null)
+			return;
+		
 		if (creature.hasBuff("co_position_secured"))
 			core.buffService.removeBuffFromCreatureByName(creature, "co_position_secured");
 		else
@@ -1318,11 +1412,11 @@ public class CombatService implements INetworkDispatch {
 
 		StartTask startTask = new StartTask(actionCounter, creature.getObjectID(), command.getCommandCRC(), CRC.StringtoCRC(command.getCooldownGroup()), command.getCooldown());
 		ObjControllerMessage objController2 = new ObjControllerMessage(0x0B, startTask);
-		creature.getClient().getSession().write(objController2.serialize());
+		client.getSession().write(objController2.serialize());
 		
 		CommandEnqueueRemove commandRemove = new CommandEnqueueRemove(creature.getObjectID(), actionCounter);
 		ObjControllerMessage objController3 = new ObjControllerMessage(0x0B, commandRemove);
-		creature.getClient().getSession().write(objController3.serialize());
+		client.getSession().write(objController3.serialize());
 
 		
 	}
@@ -1345,14 +1439,16 @@ public class CombatService implements INetworkDispatch {
 
 			@Override
 			public void run() {
-
-				if(dot.getRemainingDuration() <= 0) {
-					target.removeDot(dot);
-					dot.getTask().cancel(true);
+				try {
+					if(dot.getRemainingDuration() <= 0) {
+						target.removeDot(dot);
+						dot.getTask().cancel(true);
+					}
+					
+					doHealOverTimeTick(healer, target, command, dot);
+				} catch (Exception e) {
+					e.printStackTrace();
 				}
-				
-				doHealOverTimeTick(healer, target, command, dot);
-				
 			}
